requirements.txt
appdirs==1.4.3
Django==1.11
djangorestframework==3.6.2
facebook-sdk==2.0.0
funcsigs==1.0.2
mock==2.0.0
packaging==16.8
pbr==3.0.1
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
six==1.10.0
